openxr: Update generation of encode/decode for BaseHeaders
Properly generate the encode/decode/consumer code for more base header
scenarios.  The structures can't just be treated like the BaseHeader
struct, they need to be determined which actual structure is used,
and then encode that structure.  Similarly on decode, the appropriate
structure must be decoded, so we need to peak at what structure type
is actually in the stream and decode that, but still pass the
generic base header pointer decoder down to each consumer.  The consumers
then need to handle the appropriate actual structure as well.