openxr: Ignore invalid fences and threads
For OpenXR, we sometimes get Vulkan data from threads that are
internal to the OpenXR runtime.  Try to workaround this by
detecting "partially visible" threads and ignoring their contents.
These "partially visible" threads will trigger things like
"vkWaitForFences" or "vkResetFences" with fence values we have
never seen before.
This is because they were created inside the OpenXR runtime during
the lifetime of an OpenXR command.  Likely, they are waited on
and reset in a separate thread which is why we suddenly see them.

This includes making a setting to select this option,
"skip_threads_with_invalid_data", and this new setting is disabled
by default.