openxr: temporary fence workaround
On capture when OpenXR is enabled on a Meta Quest headset, we are
getting extraneous vkWaitForFences and vkResetFences commands that
don't line up with any of our fences.  This was tracked down to the
OpenXR runtime doing fence tracking (likely in a different thread)
outside of being wrapped in an OpenXR command (which would have already
been filtered out.
So, just do some preempting dropping of commands if we don't recognize
one ore more of the fences being used.