codegen: Update extension struct handling
This change modifies how the extension structs are stored to be more
obvious.  Previously, the extension structs were stored by the
struct that did the extending.  Instead, switch to storing the extended
structs based on the parent that is extended.

This resulted in order changes in output of the various pNext structs
so sort them so they will no longer change in order.

** Warning **
It does re-arrange items in the file, but they are all still present.
*************