Refactor FileProcessor file access
Reroute all file access though ActiveFiles::Ref
Allow for IO implementation to be centrally defined, and modified

In order support other input streams isolate the operations for a
given stream type to a mixin, selected by a using statement.  In future
we may want this build on an abstract base class.